requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
